chore: 유틸함수, 훅 분리
